Insert headers, rename files
