+++ 02_Code/fse2017_robot/controllers/LineFollowers.py
@@ -1,56 +1,50 @@
-#!/usr/bin/env python()
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+# This file is part of FSE 2017.
+#
+# FSE 2017 is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# FSE 2017 is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with FSE 2017.  If not, see <http://www.gnu.org/licenses/>.
+
+"""Line following demonstration."""
+
 import time
 from threading import Thread
 
 import RPi.GPIO as GPIO
-from HCSR04 import HCSR04  # This is the ultrasonic ranger
+from ..drivers.ultrasonic_ranger import UltrasonicRanger
 
 from ..drivers.TB6612FNG import TB6612FNG
 from ..drivers.infrared import InfraredSensor
 from ..drivers.led import LED
 
 
-class HeartBeat:
+class HeartBeat(object):
+    """Heartbeat thread."""
+
     def __init__(self):
-        self._running = True
+        self._running = False
         self._LED = LED()
 
     def terminate(self):
         self._running = False
 
     def run(self):
+        self._running = True
         while self._running:
             self._LED.on()
             time.sleep(0.5)
             self._LED.off()
             time.sleep(0.5)
-
-class AutonomousDrive:
-    def __init__(self):
-        self._running = True
-        self._USonic = HCSR04()
-        self._MotorDrive = TB6612FNG()
-
-    def terminate(self):
-        self._running = False
-
-    def run(self):
-        DutyCycle = 4
-        while self._running:
-            qualifier,distance =  self._USonic.get_averageDistance()
-            if qualifier == 1:
-                if distance < 20:
-                    self._MotorDrive.stop()
-                    time.sleep(0.01)
-                    self._MotorDrive.reverse(DutyCycle-2)
-                    time.sleep(0.2)
-                    self._MotorDrive.left(DutyCycle-2)
-                    time.sleep(0.2)
-                    self._MotorDrive.forward(DutyCycle)
-                else:
-                    self._MotorDrive.forward()
-            else:
-                pass
 
 
 class LineFollower:
@@ -66,7 +60,7 @@ class LineFollower:
         self._running = False
 
     def run(self):
-        statusLeft  = False
+        statusLeft = False
         statusRight = False
         dc = 2
         USonicDist = 1000
@@ -74,18 +68,18 @@ class LineFollower:
 
         if self._IRNbr == 2:
             while self._running:
-                #read USonic sensor
+                # read USonic sensor
 
-                #if self._loopCount%2  == True:
-                    #qualifier,dis = self._USonic.get_averageDistance()
-                    #if qualifier == 1:
-                        #USonicDist = dis
-                    #else:
-                        #print "wrong measurement"
-                #self._loopCount += 1
+                # if self._loopCount%2  == True:
+                # qualifier,dis = self._USonic.get_averageDistance()
+                # if qualifier == 1:
+                # USonicDist = dis
+                # else:
+                # print "wrong measurement"
+                # self._loopCount += 1
 
-                #read IR Sensor
-                statusLeft  = self._IR.middle
+                # read IR Sensor
+                statusLeft = self._IR.middle
                 statusRight = self._IR.right
 
                 if USonicDist > 10:
@@ -93,13 +87,13 @@ class LineFollower:
                         self._MotorDrive.forward(dc)
                     elif statusLeft:
                         # line on the left => move left
-                        self._MotorDrive.left(dc-1)
+                        self._MotorDrive.left(dc - 1)
                     elif statusRight:
                         # line on the right => move right
-                        self._MotorDrive.right(dc-1)
+                        self._MotorDrive.right(dc - 1)
                 else:
                     self._MotorDrive.stop(1)
-                    print "Obstacle detected at "+ str(USonicDist)
+                    print "Obstacle detected at " + str(USonicDist)
         else:
             print "Nooo"
 
@@ -126,11 +120,11 @@ if __name__ == "__main__":
 
     try:
         Th1 = HeartBeat()
-        HeartBeatThread = Thread(target = Th1.run)
+        HeartBeatThread = Thread(target=Th1.run)
         HeartBeatThread.start()
 
         Th2 = LineFollower()
-        RpiAutCarThread = Thread(target = Th2.run)
+        RpiAutCarThread = Thread(target=Th2.run)
         RpiAutCarThread.start()
 
         while True:
@@ -139,5 +133,3 @@ if __name__ == "__main__":
         Th1.terminate()
         Th2.terminate()
         GPIO.cleanup()
-
-
