+++ 02_Code/fse2017_robot/drivers/TB6612FNG.py
@@ -1,4 +1,20 @@
-#!/usr/bin/python
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+# This file is part of FSE 2017.
+#
+# FSE 2017 is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# FSE 2017 is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with FSE 2017.  If not, see <http://www.gnu.org/licenses/>.
+
 import time
 import sys
 import wiringpi as gpio
