--- README.md
@@ -1,12 +1,0 @@
-Rpi Autonomous Car PCB 
-====================
-This board adds these functions to a Raspberry Pi:
-- Analog Digital Converter with MCP3004 (4)
-- DC Brushed Motor driver (2) with TB6612FNG
-- Level converter (5V -> 3.3V) with TXB0108
-
-Top                                                                      |  Bottom
-:-----------------------------------------------------------------------:|:--------------------------------------------------------------:
-![Alt text](03_Docs/Components/top_gold_with_mask.png?raw=true "Top")  |  ![Alt text](03_Docs/Components/bot_gold_with_mask.png?raw=true "Bottom")
-
-Authors: Frederic Afadjigla, Daniel Lee, 2017
